+++ smartpost/__init__.py
@@ -1,4 +1,4 @@
 from smartpost.client import Client
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 __all__ = ["Client"]